+++ src/test/java/com/learning/api/controller/ReviewControlTest.java
@@ -98,7 +98,7 @@ class ReviewControlTest {
         Review review = new Review();
         review.setUserId(savedUserId);
         review.setCourseId(savedCourseId);
-        review.setRating((byte) 4);
+        review.setRating(4);
         review.setComment("Initial comment");
         savedReview = reviewRepository.save(review);
     }
@@ -227,7 +227,7 @@ class ReviewControlTest {
         Review updateBody = new Review();
         updateBody.setUserId(savedReview.getUserId());
         updateBody.setCourseId(savedReview.getCourseId());
-        updateBody.setRating((byte) 2);
+        updateBody.setRating(2);
         updateBody.setComment("Updated comment");
 
         mockMvc.perform(put("/api/reviews/{id}", savedReview.getId())
@@ -244,7 +244,7 @@ class ReviewControlTest {
         Review updateBody = new Review();
         updateBody.setUserId(savedUserId);
         updateBody.setCourseId(savedCourseId);
-        updateBody.setRating((byte) 3);
+        updateBody.setRating(3);
         updateBody.setComment("Update comment");
     
         mockMvc.perform(put("/api/reviews/{id}", 999999L)
